refactor: 변경된 파일 구조에 맞게 임포트문 수정
